% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump_gwas_results.R
\name{clump_gwas_results}
\alias{clump_gwas_results}
\title{Title}
\usage{
clump_gwas_results(
  bed.file,
  gwas.res.table,
  pval.column,
  r2thresh = 0.2,
  window.size = 1000
)
}
\arguments{
\item{bed.file}{path to bedfile, with .bed extension, need .bim and .fam in directory too}

\item{gwas.res.table}{table that contains pvalues, will be subset}

\item{pval.column}{name of pvalue column}

\item{r2thresh}{LD threshold for clumping}

\item{window.size}{size of window for clumping}
}
\value{
gwas results table subsetted to include only representative snps.
}
\description{
Title
}
\examples{
# no example
}
