% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getvarcomps.R
\name{get.varcomps}
\alias{get.varcomps}
\title{get variance components}
\usage{
get.varcomps(model.object)
}
\arguments{
\item{model.object}{a model object from a call to lmer()}
}
\value{
a data.frame of percent variance attributable to each model term
}
\description{
get variance components
}
\examples{
fm1 <- lme4::lmer(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)
get.varcomps(fm1)
}
