% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_dist_from_snp.R
\name{get_gene_dist_from_snp}
\alias{get_gene_dist_from_snp}
\title{Title get a gene's distance from a given snp}
\usage{
get_gene_dist_from_snp(s, start, end)
}
\arguments{
\item{s}{Snp physical position}

\item{start}{gene start}

\item{end}{gene end}
}
\value{
minimum distance the start or end of a gene is to the snp
}
\description{
Title get a gene's distance from a given snp
}
\examples{
# no example
}
