% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_y_pos.R
\name{get_gene_y_pos}
\alias{get_gene_y_pos}
\title{Title Generate equally spaced y values}
\usage{
get_gene_y_pos(from, to, length.out)
}
\arguments{
\item{from}{start value}

\item{to}{end value}

\item{length.out}{number of things}
}
\value{
vector of values
}
\description{
Title Generate equally spaced y values
}
\examples{
# no example
}
