% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ld_in_window.R
\name{get_ld_in_window}
\alias{get_ld_in_window}
\title{Make a table of maximum R2 of a list of snps to all snps in a window. Takes gwas output from a single gwas model.}
\usage{
get_ld_in_window(qtl.df, window, plink.path, geno.bed, in.dir, pvals.in.log)
}
\arguments{
\item{qtl.df}{table that includes only significant hits in a qtl with columns (CHR, POS, PVAL), corresponding to (chromosome, physical position, and pvalue). QTL are typically defined as hits grouped by LD by something like \code{plink --clump}}

\item{window}{kilobases on either side of top QTL snp to plot}

\item{plink.path}{path to plink 1.9 executable}

\item{geno.bed}{prefix of genotype files in plink (bed/bim/fam) format}

\item{in.dir}{directory where genotype files are located}

\item{pvals.in.log}{boolean, are pvalues in input data.frames in -log10(p)?}
}
\value{
Named list with 2 items
\itemize{
\item table: table with marker.IDs (CHR-POS) and maximum LD in R2 for each snp to the snps in the qtl.df
\item key.snp: marker.ID corresponding to the middle of the window
}
}
\description{
Make a table of maximum R2 of a list of snps to all snps in a window. Takes gwas output from a single gwas model.
}
\examples{
# Work in progress
}
