% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ld_in_window_multi.R
\name{get_ld_in_window_multi}
\alias{get_ld_in_window_multi}
\title{Make a table of maximum R2 of a list of snps to all snps in a window. Takes gwas output from multiple gwas models.}
\usage{
get_ld_in_window_multi(
  qtl.df,
  window,
  plink.path,
  geno.bed,
  in.dir,
  pvals.in.log,
  grouping.variable,
  top.only = T
)
}
\arguments{
\item{qtl.df}{table that includes only significant hits in a qtl with columns (CHR, POS, PVAL), corresponding to (chromosome, physical position, and pvalue). QTL are typically defined as hits grouped by LD by something like \code{plink --clump}}

\item{window}{kilobases on either side of top QTL snp to plot}

\item{plink.path}{path to plink 1.9 executable}

\item{geno.bed}{prefix of genotype files in plink (bed/bim/fam) format}

\item{in.dir}{directory where genotype files are located}

\item{pvals.in.log}{boolean, are pvalues in input data.frames in -log10(p)?}

\item{grouping.variable}{string indicating column name that corresponds to unique gwas models in the qtl.df}

\item{top.only}{calculate LD for top p-value snps in each grouping.variable (T) or all snps in qtl.df (F)}
}
\value{
Named list with 2 items
\itemize{
\item table: table with marker.IDs (CHR-POS) and maximum LD (R2) for each snp in window to the snps (top or all) in the qtl.df
\item key.snp: marker.ID corresponding to the middle of the window
}
}
\description{
Make a table of maximum R2 of a list of snps to all snps in a window. Takes gwas output from multiple gwas models.
}
\examples{
# Work in progress
}
