% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rep_snps_gwas.R
\name{get_rep_snps_gwas}
\alias{get_rep_snps_gwas}
\title{Get snps that represent LD blocks using pvalues}
\usage{
get_rep_snps_gwas(
  bed.file,
  gwas.res.table,
  pval.column,
  r2thresh = 0.2,
  window.size = 1000
)
}
\arguments{
\item{bed.file}{path to bedfile, with .bed extension, need .bim and .fam in directory too}

\item{gwas.res.table}{table that contains pvalues, will be subset}

\item{pval.column}{name of pvalue column}

\item{r2thresh}{LD threshold for clumping}

\item{window.size}{size of window for clumping}
}
\value{
gwas results table subsetted to include only representative snps.
}
\description{
Get snps that represent LD blocks using pvalues
}
\examples{
# no example
}
