% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_zoom_multi.R
\name{locus_zoom_multi}
\alias{locus_zoom_multi}
\title{Make a locus zoom style plot with points colored by R2 to a qtl and nearby genes annotated. Takes gwas output from multiple gwas models.}
\usage{
locus_zoom_multi(
  gwas.res,
  qtl.df,
  annotation.table,
  pvals.in.log = T,
  plink.path,
  geno.bed.filename,
  geno.bed.directory = "/.",
  plot.r2.thresh = 0.2,
  window,
  sig.line,
  plot.title = "",
  unique.gwas.model.variable,
  sig.hit.color.var = NULL,
  sig.hit.shape.var = NULL,
  shape.scale = NULL,
  include.gene.id = FALSE,
  plot.effect = TRUE
)
}
\arguments{
\item{gwas.res}{table of all gwas results, should contain columns (\code{CHR}, \code{POS}, \code{PVAL}), corresponding to (chromosome, physical position, and pvalue).}

\item{qtl.df}{table with same columns that includes only significant hits in a qtl. QTL are typically defined as hits grouped by LD by something like \code{plink --clump}.}

\item{annotation.table}{table with annotations with columns (geneID, CHR, start, end, annotation). start and end correspond to base-pair coordinates of start and end of gene. CHR is chromosome of gene.}

\item{pvals.in.log}{boolean, are pvalues in input data.frames in -log10(p) format?}

\item{plink.path}{path to plink 1.9 executable.}

\item{geno.bed.filename}{prefix of genotype files in plink (bed/bim/fam) format.}

\item{geno.bed.directory}{directory where genotype files are located.}

\item{plot.r2.thresh}{minimum LD with qtl snps to plot snps colored by LD. SNPs below are plotted in grey.}

\item{window}{kilobases on either side of top QTL snp to plot.}

\item{sig.line}{-log10(p) value to draw line on plot.}

\item{plot.title}{Optional. Title of plot.}

\item{unique.gwas.model.variable}{string indicating column name in gwas.res and qtl.df that corresponds to unique gwas models (e.g. phenotype within a location). See \link{get_ld_in_window_multi} for usage in LD calculation.}

\item{sig.hit.color.var}{Optional. String indicating a column in gwas.res and qtl.df that will color significant hits. (e.g. a phenotype or location)}

\item{sig.hit.shape.var}{Optional. String indicating a column in gwas.res and qtl.df that will change shape of significant hits. (e.g. a phenotype or location)}

\item{shape.scale}{Optional. A defined shape scale to keep plotting of shapes consistent when all factors are note present. See another great function I made to do this.}

\item{include.gene.id}{boolean, include geneID in gene annotations or not}

\item{plot.effect}{boolean, include plot of pvalues vs effect size? If TRUE, \code{gwas.res} and \code{qtl.df} must have column \code{EFF} that conatins gwas effect sizes.}

\item{geno.bed}{prefix of genotype files in plink (bed/bim/fam) format.}
}
\value{
GGplot of locus zoom style plot with points colored by maximum R2 to markers in the qtl.df. Only results from \code{unique.gwas.model.variable} present in the qtl are plotted. Alongside are annotations of nearby genes.
}
\description{
Make a locus zoom style plot with points colored by R2 to a qtl and nearby genes annotated. Takes gwas output from multiple gwas models.
}
\examples{
# work in progress
}
