% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_zoom_single.R
\name{locus_zoom_single}
\alias{locus_zoom_single}
\title{Make a locus zoom style plot with results from a single gwas, points colored by R2 to a qtl and nearby genes annotated.}
\usage{
locus_zoom_single(
  gwas.res,
  qtl.df,
  annotation.table,
  pvals.in.log = T,
  plink.path,
  geno.bed.filename,
  geno.bed.directory = "/.",
  plot.r2.thresh = 0.2,
  window,
  sig.line,
  plot.title = "",
  include.gene.id = F,
  plot.effect = T
)
}
\arguments{
\item{gwas.res}{table of all gwas results, should contain columns (\code{CHR}, \code{POS}, \code{PVAL}), corresponding to (chromosome, physical position, and pvalue).}

\item{qtl.df}{table with same columns that includes only significant hits in a qtl. QTL are typically defined as hits grouped by LD by something like \code{plink --clump}.}

\item{annotation.table}{table with annotations with columns (geneID, CHR, start, end, annotation). start and end correspond to base-pair coordinates of start and end of gene. CHR is chromosome of gene.}

\item{pvals.in.log}{boolean, are pvalues in input data.frames in -log10(p) format?}

\item{plink.path}{path to plink 1.9 executable.}

\item{geno.bed.filename}{prefix of genotype files in plink (bed/bim/fam) format.}

\item{geno.bed.directory}{directory where genotype files are located.}

\item{plot.r2.thresh}{minimum LD with qtl snps to plot snps colored by LD. SNPs below are plotted in grey.}

\item{window}{kilobases on either side of top QTL snp to plot.}

\item{sig.line}{-log10(p) value to draw line on plot.}

\item{plot.title}{Optional. Title of plot.}

\item{include.gene.id}{boolean, include geneID in gene annotations or not}

\item{plot.effect}{boolean, include plot of pvalues vs effect size? If TRUE, \code{gwas.res} and \code{qtl.df} must have column \code{EFF} that conatins gwas effect sizes.}

\item{geno.bed}{prefix of genotype files in plink (bed/bim/fam) format.}
}
\value{
ggplot object of plot
}
\description{
Make a locus zoom style plot with results from a single gwas, points colored by R2 to a qtl and nearby genes annotated.
}
\examples{
# work in progress
}
