% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luebbert-package.R
\docType{package}
\name{luebbert-package}
\alias{luebbert}
\alias{luebbert-package}
\title{luebbert: Some helper functions that I commonly use}
\description{
This is a personal packages, a way to source functions that I use often. These are things like ggplot tweaks and smaller functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cluebbert/luebbert}
  \item Report bugs at \url{https://github.com/cluebbert/luebbert/issues}
}

}
\author{
\strong{Maintainer}: Collin Luebbert \email{cluebbert@danforthcenter.org} (\href{https://orcid.org/0000-0002-8665-5950}{ORCID})

}
\keyword{internal}
