% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_consistent_scale.R
\name{make.consistent.scale}
\alias{make.consistent.scale}
\title{Make a ggplot scale that has consistent correspondence between aesthetics and variables}
\usage{
make.consistent.scale(
  values,
  vars,
  type = c("fill", "color", "shape"),
  show.example = F,
  name = NULL
)
}
\arguments{
\item{values}{vector, values to supply to given scale (e.g. colors or shapes)}

\item{vars}{character, variables to link to values. Will have one to one correspondence to values based on order}

\item{type}{character, either "fill", "color" or "shape"}

\item{show.example}{boolean, print a plot to show what values are tied to which variables}

\item{name}{optional, string to name scale, will appear as legend heading when used in a ggplot}
}
\value{
ggplot scale object for use in downstream plotting
}
\description{
Make a ggplot scale that has consistent correspondence between aesthetics and variables
}
\examples{
# work in progress
}
