% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clumps.R
\name{make_clumps}
\alias{make_clumps}
\title{Assign snps to clumps based on LD after gwas}
\usage{
make_clumps(
  geno.bed,
  gwas.res,
  pvals.in.log = T,
  window = 500,
  ld.thresh = 0.5
)
}
\arguments{
\item{geno.bed}{character, bedfile path, no .bed extension}

\item{gwas.res}{data.frame, table of gwas results with columns (marker.ID, CHR, POS, PVAL)}

\item{pvals.in.log}{boolean, are pvalues in \code{gwas.res} in -log10(p) or not}

\item{window}{integer, window in kilobases either side of snp to look for snps in LD,}

\item{ld.thresh}{numeric, R2 threshold above which snps will be grouped}
}
\value{
table with columns marker.ID and clump_num. Clump_num indicates groupings, numberings start from the larges pvalue to smallest. May want to reassign afterwards to be along the genome.
}
\description{
Assign snps to clumps based on LD after gwas
}
\examples{
# work in progress
}
