% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_effect_plot.R
\name{make_effect_plot}
\alias{make_effect_plot}
\title{Make volcano style effect size vs pvalue plot}
\usage{
make_effect_plot(
  gwas.res,
  qtl.df,
  pvals.in.log = TRUE,
  plot.r2.thresh = 0.2,
  ld.list,
  window,
  sig.line,
  include.legend = FALSE
)
}
\arguments{
\item{gwas.res}{table of all gwas results, should contain columns (CHR, POS, PVAL), corresponding to (chromosome, physical position, and pvalue).}

\item{qtl.df}{table with same columns that includes only significant hits in a qtl. QTL are typically defined as hits grouped by LD by something like \code{plink --clump}}

\item{pvals.in.log}{boolean, are pvalues in input data.frames in -log10(p)?}

\item{plot.r2.thresh}{minimum LD with qtl snps to plot snps colored by LD}

\item{ld.list}{output of \link{get_ld_in_window}}

\item{window}{kilobases on either side of top QTL snp to plot}

\item{sig.line}{-log10(p) value to draw line on plot}

\item{include.legend}{include a legend in the plot}
}
\value{
GGplot object of plot. Points colored by maximum R2 to snps in qtl.df
}
\description{
Make volcano style effect size vs pvalue plot
}
\examples{
# Work in progress
}
