% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ld.R
\name{make_ld}
\alias{make_ld}
\title{Title Make a table of LD around a certain snp}
\usage{
make_ld(snp.name, window = 1000, bedfile, in.dir, out.filename = "ld_out_temp")
}
\arguments{
\item{snp.name}{Name of snp around which to calculate LD}

\item{window}{Window around snp in KB}

\item{bedfile}{bedfile prefix, no .bed extension}

\item{in.dir}{where the bedfile lives}

\item{out.filename}{name of outfile, will be placed in in.dir}
}
\value{
outputs table of LD, for use within plotting function mostly
}
\description{
Title Make a table of LD around a certain snp
}
\examples{
# No example
}
