% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_manhattan.R
\name{make_manhattan}
\alias{make_manhattan}
\title{Make a manhattan plot from table of gwas results}
\usage{
make_manhattan(
  df,
  chrom.col,
  pos.col,
  pval.col,
  pval.in.log = F,
  bonf.cutoff,
  smoothing.cutoff,
  highlight.df = NULL
)
}
\arguments{
\item{df}{data.frame with chromosome, position and pvalue columns}

\item{chrom.col}{character - name of chromosome column}

\item{pos.col}{character - name of position column}

\item{pval.col}{character - name of pvalue column}

\item{pval.in.log}{boolean - is pvalue column log10 or not}

\item{bonf.cutoff}{significance cutoff}

\item{smoothing.cutoff}{minimum pvalue to plot to reduce size of plot}

\item{highlight.df}{optional - dataframe in same format as input that has snps to highlight}
}
\value{
manhattan plot
}
\description{
Make a manhattan plot from table of gwas results
}
