% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_manhattan_offsets.R
\name{make_manhattan_offsets}
\alias{make_manhattan_offsets}
\title{Make table of offsets for plotting manhattans}
\usage{
make_manhattan_offsets(df, chrom.col, pos.col)
}
\arguments{
\item{df}{dataframe of gwas results with minimally chromosome and position columns}

\item{chrom.col}{character - name of chromosome column}

\item{pos.col}{character - name of position column}
}
\value{
table of offsets to add for each chromosome to get genome in continuous line
}
\description{
Make table of offsets for plotting manhattans
}
