% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_manhattan_zoom_annotation.R
\name{make_manhattan_zoom_annotation}
\alias{make_manhattan_zoom_annotation}
\title{Title Make a manhattan in a region and label genes and LD for a table of top gwas hits}
\usage{
make_manhattan_zoom_annotation(
  out.dir = "./",
  plot.type = c("png", "pdf"),
  trait.name,
  top.hits,
  gwas.res,
  window = 1000,
  geno.bed,
  plot.r2.thresh,
  file.prefix = "",
  annotation.table
)
}
\arguments{
\item{out.dir}{where to output plots}

\item{plot.type}{png or pdf?}

\item{trait.name}{character string of name of trait}

\item{top.hits}{table of top hits with columns CHROM, POS, and PVAL column, p-value is log transformed in function so should not be pre-transformed}

\item{gwas.res}{table of all gwas results, whats plotted}

\item{window}{window in KB to plot in}

\item{geno.bed}{prefix of bed file, character string}

\item{plot.r2.thresh}{minimum LD to plot}

\item{file.prefix}{optional label for output files}

\item{annotation.table}{table with gene descriptions and start/stop, currently only supports a specific table from maize}
}
\value{
outputs 1 plot per row of top.hits table
}
\description{
Title Make a manhattan in a region and label genes and LD for a table of top gwas hits
}
\examples{
# no example
}
