% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sideways_manhattan.R
\name{make_sideways_manhattan}
\alias{make_sideways_manhattan}
\title{Make sideways manhattan plot for building locus zoom. Receives output from a single gwas model.}
\usage{
make_sideways_manhattan(
  gwas.res,
  qtl.df,
  pvals.in.log = TRUE,
  plot.r2.thresh = 0.2,
  ld.list,
  window,
  sig.line
)
}
\arguments{
\item{gwas.res}{table of all gwas results, should contain columns (CHR, POS, PVAL), corresponding to (chromosome, physical position, and pvalue).}

\item{qtl.df}{table with same columns that includes only significant hits in a qtl. QTL are typically defined as hits grouped by LD by something like \code{plink --clump}}

\item{pvals.in.log}{boolean, are pvalues in input data.frames in -log10(p)?}

\item{plot.r2.thresh}{minimum LD with qtl snps to plot snps colored by LD}

\item{ld.list}{output of \link{get_ld_in_window}}

\item{window}{kilobases on either side of top QTL snp to plot}

\item{sig.line}{-log10(p) value to draw line on plot}
}
\value{
GGplot of manhattan plot with points colored by maximum R2 to markers in the qtl.df.
}
\description{
Make sideways manhattan plot for building locus zoom. Receives output from a single gwas model.
}
\examples{
# Work in progress
}
