% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sideways_manhattan_multi.R
\name{make_sideways_manhattan_multi}
\alias{make_sideways_manhattan_multi}
\title{Make sideways manhattan plot for building locus zoom. Receives output from multiple gwas models.}
\usage{
make_sideways_manhattan_multi(
  gwas.res,
  qtl.df,
  pvals.in.log = TRUE,
  plot.r2.thresh = 0.2,
  ld.list,
  window,
  sig.line,
  unique.gwas.model.variable = NULL,
  sig.hit.color.var = NULL,
  sig.hit.shape.var = NULL,
  shape.scale = NULL
)
}
\arguments{
\item{gwas.res}{table of all gwas results, should contain columns (CHR, POS, PVAL), corresponding to (chromosome, physical position, and pvalue).}

\item{qtl.df}{table with same columns that includes only significant hits in a qtl. QTL are typically defined as hits grouped by LD by something like \code{plink --clump}}

\item{pvals.in.log}{boolean, are pvalues in input data.frames in -log10(p)?}

\item{plot.r2.thresh}{minimum LD with qtl snps to plot snps colored by LD. SNPs below are plotted in grey.}

\item{ld.list}{output of \link{get_ld_in_window_multi}}

\item{window}{kilobases on either side of top QTL snp to plot}

\item{sig.line}{-log10(p) value to draw line on plot}

\item{unique.gwas.model.variable}{string indicating column name in gwas.res and qtl.df that corresponds to unique gwas models (e.g. phenotype within a location). See \link[luebbert]{get_ld_in_window_multi} for usage in LD calculation.}

\item{sig.hit.color.var}{Optional. String indicating a column in gwas.res and qtl.df that will color significant hits. (e.g. a phenotype or location)}

\item{sig.hit.shape.var}{Optional. String indicating a column in gwas.res and qtl.df that will change shape of significant hits. (e.g. a phenotype or location)}

\item{shape.scale}{Optional. A defined shape scale to keep plotting of shapes consistent when all factors are note present. See another great function I made to do this.}
}
\value{
GGplot of manhattan plot with points colored by maximum R2 to markers in the qtl.df. Only results from gwas models present in the qtl are plotted.
}
\description{
Make sideways manhattan plot for building locus zoom. Receives output from multiple gwas models.
}
\examples{
# Work in progress
}
