% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pc_contribs.R
\name{plot_pc_contribs}
\alias{plot_pc_contribs}
\title{Plot PC contributions}
\usage{
plot_pc_contribs(pca_obj, plot.title = "", num.pcs = 6, label.offset)
}
\arguments{
\item{pca_obj}{pca object created from "princomp()"}

\item{plot.title}{title for plot}

\item{num.pcs}{number pcs to plot either as maximum or as vector of desired pc's}

\item{label.offset}{percentage to offset percent contribution label from side of plot}
}
\value{
GGplot of variable contributions for each PC
}
\description{
Plot PC contributions
}
\examples{
#pcobj <- princomp(data)
#plot_pc_contribs(pcobj)
}
