% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rite_text_file.R
\name{rite_text_file}
\alias{rite_text_file}
\title{make text file with correct encoding for use on linux system}
\usage{
rite_text_file(filename, data)
}
\arguments{
\item{filename}{name and path of output file (without trailing .txt)}

\item{data}{vector to be output to text file}
}
\value{
vector as text file, each element on separate line
}
\description{
make text file with correct encoding for use on linux system
}
\examples{
# no example

}
