% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ritecsv.R
\name{ritecsv}
\alias{ritecsv}
\title{write.csv with row.names = F as default}
\usage{
ritecsv(..., row.names = FALSE)
}
\arguments{
\item{...}{extra options to base r write.csv() function}

\item{row.names}{save rownames or not, F is default}
}
\value{
saves csv with no rownames by default
}
\description{
write.csv with row.names = F as default
}
