% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ld.R
\name{make_ld}
\alias{make_ld}
\title{Make a table of LD around a certain snp}
\usage{
make_ld(
  plink.path = "~/bin/plink",
  snp.name,
  window,
  bedfile,
  in.dir,
  out.filepath = "ld_out_temp"
)
}
\arguments{
\item{plink.path}{path to plink 1.9 executable}

\item{snp.name}{Name of snp around which to calculate LD}

\item{window}{Window around snp in KB}

\item{bedfile}{bedfile path, no .bed extension}

\item{in.dir}{directory of bedfile, include trailing "/"}

\item{out.filepath}{name and path of outfile}
}
\value{
outputs table of LD, for use within plotting function mostly
}
\description{
Make a table of LD around a certain snp
}
\examples{
# No example
}
