% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ld.R
\name{make_ld}
\alias{make_ld}
\title{Title Make a table of LD around a certain snp}
\usage{
make_ld(snp.name, window = 1000, bedfile, out.filepath = "./ld_out_temp")
}
\arguments{
\item{snp.name}{Name of snp around which to calculate LD}

\item{window}{Window around snp in KB}

\item{bedfile}{bedfile path, no .bed extension}

\item{out.filepath}{path of outfile}
}
\value{
outputs table of LD, for use within plotting function mostly
}
\description{
Title Make a table of LD around a certain snp
}
\examples{
# No example
}
