% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycorrplot.R
\name{mycorrplot}
\alias{mycorrplot}
\title{Make visuzlization of correlation matrix}
\usage{
mycorrplot(
  dat,
  use = "pairwise.complete",
  axis.text.size = 1,
  add.coef = FALSE,
  coef.text.size = 1,
  color.scheme = c("sunset", "BuRd", "nightfall", "PRGn"),
  out.file.png = NULL,
  pvalthresh = 0.05,
  ...
)
}
\arguments{
\item{dat}{dataframe of data to compute correlations, columns are traits}

\item{use}{how to deal with missing data in correlation, default is pairwise complete}

\item{axis.text.size}{size of trait labels in plot}

\item{add.coef}{plot coefficient or not}

\item{coef.text.size}{size of correlation coefficient labels in plot}

\item{color.scheme}{which color scheme to use, diverging schemes from khroma package}

\item{out.file.png}{where to output png file}

\item{pvalthresh}{threshold for significance}

\item{...}{extra options passed to corrplot::corrplot}
}
\value{
prints visualization of correlation plot
}
\description{
Make visuzlization of correlation matrix
}
